Backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.1
python-dotenv==1.0.0
werkzeug==2.3.7
flask-cors==3.0.10